Inline frequent methods.
